Implement child nodes with an Array instead of a List.
This changes reduces memory usage on the benchmark from 959.6MB to
775.2MB (tested on Window x64). This is due to the fact that List has some
overhead to support growing, in particular it has a second level of indirection
as List is a class that holds a reference to an Array. This comes at the price
of allocating a completely new Array when growing a node as terms are added.
This cost is not significant however as the benchmark shows that the time taken
to create the trie actually goes down from 18.8ms to 15.3ms. This is probably
dues to the fact that the data set is sparse and most nodes are not grown to the
full size of the alphabet. Finally, the reduced pointer chasing also improves
lookup time by around 20%.